+++ src/Formats/Mobi/MobiModule.php
@@ -44,7 +44,7 @@ class MobiModule extends EbookModule
             $this->ebook->setAuthor(new BookAuthor($author));
         }
 
-        foreach ($this->parser->get(MobiReader::ISBN_104, true) as $isbn) {
+        foreach ($this->parser->get(MobiReader::ISBN_104, true) ?? [] as $isbn) {
             $this->ebook->setIdentifier(new BookIdentifier($isbn));
         }
 
@@ -61,7 +61,7 @@ class MobiModule extends EbookModule
         $description = $this->parser->get(MobiReader::DESCRIPTION_103);
         $this->ebook->setDescription($description);
 
-        foreach ($this->parser->get(MobiReader::SUBJECT_105, true) as $value) {
+        foreach ($this->parser->get(MobiReader::SUBJECT_105, true) ?? [] as $value) {
             $this->ebook->setTag($value);
         }